Refactor: Convert many C style casts to C++ style and id_cast
For improved type safety and to follow code style guidelines, this is a
largely automated conversion of code in source/ and intern/cycles/blender.

* Use const_cast and static_cast where possible and reintepret_cast if not.
* For arithmetic and enumeration types, use functional notation.
* Prefer blender::id_cast where possible. This required some small code
  tweaks to avoid doing this cast before the ID type has been checked. There
  are probably a few more like this lurking that will trigger asserts but not
  affect release builds.
* Change Object.data to be ID* rather than void* to take more advantage of
  id_cast checks.

Not all casts where converted, only the ones that were easy to automate.
Various modules in intern/ would also benefit but are not handled as part of
this.

Pull Request: https://projects.blender.org/blender/blender/pulls/152327

+++ source/blender/gpu/metal/mtl_pso_descriptor_state.hh
@@ -119,7 +119,7 @@ struct MTLVertexDescriptor {
 
   uint64_t hash() const
   {
-    uint64_t hash = (uint64_t)(this->max_attribute_value ^ this->num_vert_buffers);
+    uint64_t hash = uint64_t(this->max_attribute_value ^ this->num_vert_buffers);
     for (const int a : IndexRange(this->max_attribute_value + 1)) {
       hash ^= this->attributes[a].hash() << a;
     }
@@ -253,8 +253,7 @@ struct MTLRenderPipelineStateDescriptor {
     hash ^= uint64_t(this->num_color_attachments) << 16;     /* up to 6 (3 bits). */
     hash ^= uint64_t(this->depth_attachment_format) << 18;   /* up to 555 (9 bits). */
     hash ^= uint64_t(this->stencil_attachment_format) << 20; /* up to 555 (9 bits). */
-    hash ^= uint64_t(
-        *((uint64_t *)&this->vertex_descriptor.prim_topology_class)); /* Up to 3 (2 bits). */
+    hash ^= (*((uint64_t *)&this->vertex_descriptor.prim_topology_class)); /* Up to 3 (2 bits). */
 
     /* Only include elements in Hash if they are needed - avoids variable null assignments
      * influencing hash. */
